backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
Flask-Bcrypt==1.0.1
PyMySQL==1.1.0
python-dotenv==1.0.0
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
Werkzeug==2.3.7
PyJWT==2.8.0
cryptography==41.0.7
requests==2.31.0
gunicorn==21.2.0
python-dateutil==2.8.2 